Add files via upload
The project is to make a railway ticket booking system.  The GUI is designed using QtDesigner.The final code for the project is present in the loginf.py file. The UI files have also been uploaded along with the project which can be opened with the QtDesigner.exe in the python folder.